separe post test in 2 files
